Update upload_template docstring, and back out some recent changes.
For some dumb reason I had updated the behavior of upload_template so that it
no longer worked as advertised. In retrospect it really should work as
advertised, namely that the "destination" string is given verbatim to the call
to "cp".

This gives more power to the user, and the user may explicitly tack on
a trailing slash if they want "cp" to copy a file into a folder, as normal.

Signed-off-by: PlanarPlatypus <github@finalhaven.org>