Fix race condition in run/sudo.
See b9a7c9df for how the problem was (at least most recently) introduced.

Since this change here is not quite a revert of that earlier commit (note that
we kept the recv_exit_status check *before* the thread joins) I am hoping that
this change does not introduce more bugs of its own; this part of the code has
a very long history of back-and-forthing :(

However, I verified that the race condition definitely existed beforehand
(Fab task resulting in single line of output was run a few dozen times
in a row; almost half of the time the line never showed up) and that this
change fixes it (now every single invocation results in the expected output).

Thanks to Evan Jones for catching this and suggesting the move of
channel.close().

Signed-off-by: PlanarPlatypus <github@finalhaven.org>